requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
python-dotenv==1.0.1
streamlit==1.38.0
pydantic==2.9.2
python-multipart==0.0.9
faster-whisper
pyttsx3
webrtcvad-wheels
scipy
soundfile
audio-recorder-streamlit
